refactor: rename go_bot.py -> bot.py
