+++ deeppavlov/skills/go_bot/bot.py
